fix(den-web): reposition runtime version
