+++ ee/apps/den-web/app/(den)/dashboard/_components/org-dashboard-shell.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { useEffect, useMemo, useState } from "react";
+import { useMemo, useState } from "react";
 import {
   BarChart3,
   ChevronDown,
@@ -165,60 +165,6 @@ export function SidebarBrandMark({
       />
     </div>
   );
-}
-
-export function parseDenRuntimeVersion(value: unknown) {
-  if (typeof value !== "object" || value === null) {
-    return null;
-  }
-
-  const version = Object.getOwnPropertyDescriptor(value, "version")?.value;
-  return typeof version === "string" && version.trim() ? version.trim() : null;
-}
-
-export function DenRuntimeVersionLabel({ version }: { version: string }) {
-  return (
-    <p
-      className="mb-2 px-2 text-[10px] font-medium tabular-nums text-gray-400"
-      data-den-runtime-version={version}
-      title={`Den API version ${version}`}
-    >
-      Den {version}
-    </p>
-  );
-}
-
-function DenRuntimeVersion() {
-  const [version, setVersion] = useState<string | null>(null);
-
-  useEffect(() => {
-    const controller = new AbortController();
-
-    async function loadVersion() {
-      try {
-        const response = await fetch("/api/den/health", {
-          cache: "no-store",
-          signal: controller.signal,
-        });
-        if (!response.ok) {
-          return;
-        }
-
-        const nextVersion = parseDenRuntimeVersion(await response.json());
-        if (!controller.signal.aborted) {
-          setVersion(nextVersion);
-        }
-      } catch {
-        // Version metadata is optional UI; the dashboard remains quiet when
-        // den-api is unavailable or an older deployment omits the field.
-      }
-    }
-
-    void loadVersion();
-    return () => controller.abort();
-  }, []);
-
-  return version ? <DenRuntimeVersionLabel version={version} /> : null;
 }
 
 function getDashboardPageTitle(pathname: string, orgSlug: string | null) {
@@ -679,8 +625,6 @@ export function OrgDashboardShell({ children }: { children: React.ReactNode }) {
       </nav>
 
       <div className="mt-auto p-3">
-        <DenRuntimeVersion />
-
         {orgSwitcher}
 
         {orgBusy ? (
