+++ evals/flows/den-api-runtime-version.flow.mjs
@@ -17,19 +17,19 @@ function witness(ctx, condition, assertion, actual) {
 
 export default {
   id: FLOW_ID,
-  title: "Den Web discreetly shows the running Den version",
+  title: "Org settings shows the running Den version inline",
   kind: "user-facing",
   preserveTheme: true,
   requiredEnv: ["OPENWORK_EVAL_DEN_API_URL", "OPENWORK_EVAL_DEN_WEB_URL", "OPENWORK_EVAL_DEN_API_VERSION"],
   steps: [
     {
-      name: "The dashboard sidebar shows the running Den version",
+      name: "Org settings shows the running Den version beside its description",
       run: async (ctx) => {
         const apiBaseUrl = ctx.env.OPENWORK_EVAL_DEN_API_URL.trim().replace(/\/+$/, "");
         const webBaseUrl = ctx.env.OPENWORK_EVAL_DEN_WEB_URL.trim().replace(/\/+$/, "");
         const expectedVersion = ctx.env.OPENWORK_EVAL_DEN_API_VERSION.trim();
 
-        await ctx.prove("An operator can discreetly see the running Den version in the dashboard sidebar", {
+        await ctx.prove("An operator can see the running Den version beside the Org settings description", {
           voiceover: vo[0],
           action: async () => {
             await ctx.client.send("Emulation.setDeviceMetricsOverride", {
@@ -63,9 +63,14 @@ export default {
               timeoutMs: 45_000,
               label: "signed-in Den dashboard",
             });
+            await ctx.client.send("Page.navigate", { url: `${webBaseUrl}/dashboard/org-settings` });
+            await ctx.waitFor("location.pathname === '/dashboard/org-settings'", {
+              timeoutMs: 30_000,
+              label: "Org settings page",
+            });
             await ctx.waitFor(`Boolean(document.querySelector('[data-den-runtime-version=${JSON.stringify(expectedVersion)}]'))`, {
               timeoutMs: 20_000,
-              label: "Den runtime version in dashboard sidebar",
+              label: "Den runtime version beside the Org settings description",
             });
           },
           assert: async () => {
@@ -78,21 +83,27 @@ export default {
               if (!element) return null;
               const style = getComputedStyle(element);
               const bounds = element.getBoundingClientRect();
+              const description = element.previousElementSibling;
+              const descriptionBounds = description?.getBoundingClientRect();
               return {
                 text: element.textContent?.trim() ?? '',
                 version: element.getAttribute('data-den-runtime-version'),
-                fontSize: style.fontSize,
+                color: style.color,
+                parentDisplay: getComputedStyle(element.parentElement).display,
                 left: Math.round(bounds.left),
-                bottom: Math.round(bounds.bottom),
-                viewportHeight: window.innerHeight,
+                top: Math.round(bounds.top),
+                descriptionText: description?.textContent?.trim() ?? '',
+                descriptionRight: descriptionBounds ? Math.round(descriptionBounds.right) : null,
+                descriptionTop: descriptionBounds ? Math.round(descriptionBounds.top) : null,
               };
             })()`);
-            witness(ctx, label?.text === `Den ${expectedVersion}` && label?.version === expectedVersion, "The sidebar label reports the running Den version", label);
-            witness(ctx, label?.fontSize === "10px" && label?.left < 260 && label?.bottom <= label?.viewportHeight, "The version remains discreetly positioned at the bottom of the sidebar", label);
+            witness(ctx, label?.text === `Den ${expectedVersion}` && label?.version === expectedVersion, "The inline label reports the running Den version", label);
+            witness(ctx, label?.descriptionText === "Control your organization's settings." && label?.parentDisplay === "inline-flex" && label?.left > label?.descriptionRight && Math.abs(label?.top - label?.descriptionTop) <= 2, "The version sits directly to the right of the Org settings description", label);
+            witness(ctx, label?.color === "rgb(156, 163, 175)", "The version uses the light-gray metadata color", label);
           },
           screenshot: {
-            name: "den-web-runtime-version",
-            requireText: ["Dashboard", `Den ${expectedVersion}`],
+            name: "den-web-org-settings-runtime-version",
+            requireText: ["Org settings", "Control your organization's settings.", `Den ${expectedVersion}`],
           },
         });
       },
