+++ evals/voiceovers/den-api-runtime-version.md
@@ -1,3 +1,3 @@
 # Den API runtime version
 
-1. At the bottom of the Den dashboard sidebar, the running Den version stays available without competing with the workspace controls or navigation.
+1. Org settings now puts the running Den version directly beside the page description in light gray, keeping the build identity visible without adding another control.
